pyproject.toml
[project]
name = "dontforget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
